backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==1.4.53
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI Integration
openai==1.107.1

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS and middleware
fastapi-cors==0.0.6

# Email validation
email-validator==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
